# app/public/docs/api.yaml
openapi: "3.0.3"
info:
  title: MagicWebb API
  version: "1.0.0"
  description: |
    MagicWebb is a decentralized NFT marketplace on Flare Coston2.
    All `/api/v1/*` routes are rate-limited (60 req/min per IP).
    Routes marked 🔒 require a valid SIWE JWT (Bearer token or `mw_s_*` cookie).
servers:
  - url: https://magicwebb.fly.dev
    description: Production (Flare Coston2 testnet)
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/v1/wallet/{addr}/nfts:
    get:
      summary: List NFTs owned by address
      description: Returns all ERC-721 and ERC-1155 tokens owned by address, newest first.
      tags: [Wallet]
      parameters:
        - name: addr
          in: path
          required: true
          schema: { type: string, example: "0x..." }
      responses:
        "200":
          description: Array of owned NFTs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OwnedNFT"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Notifications (🔒 JWT) ──────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: "SIWE JWT via Authorization: Bearer <token>"
    CookieAuth:
      type: apiKey
      in: cookie
      name: mw_s_
      description: "HttpOnly session cookie with addr prefix (e.g. mw_s_0xabcd...)"
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "internal error" }
  schemas:
...
    OwnedNFT:
      type: object
      properties:
        collection: { type: string }
        token_id:   { type: string }
        units:      { type: string }
        standard:   { type: string, enum: [erc721, erc1155] }
        name:       { type: string }
        image_uri:  { type: string }

    # ── Notifications ─────────────────────────────────────────────────
...